Use a discarding writer for klog output
Configure klog to write messages to a discarding writer instead of
writing them to /dev/null.

The change does not affect klog behavior for the stderrthreshold flag.

It fixes crashes when runing on a platform without /dev/null device (for
example on Windows). The klog will exit the program if /dev/null file
cannot be open.

Fixes derailed/k9#703