Use ReadyColIndex to get the stable index
This relies on the newly added function ReadyColIndex which gets the
column index of the READY column in a reliable way instead of assuming
it is following the NAME field. It fixes the crash when the name is not
in the expected index when custom views are used.
It closes #1153.