feat: change sort col using two fixed keys
Change the sort col using two fixed keys instead of
using the fixed bindings per column. Another fixed key is used
to change the sort order. The behaviour is now similar to `top`

The implementation is based on re-evaluating which columns of the
model should be visible, instead of checking the currently displayed
columns. The displayed column names might be decorated with extra
content which make string comparisons cumbersome.

Also, fix the sorting on CAPACITY-like columns (e.g. pv, pvc) by
comparing the values as a `Quantity` and not `strings`.