Log toggles: add a space after "On" in logs view
Toggling any of the log commands on and off causes the log indicator to
dance around - I assume this is because the line is centered, and when
on it is N characters long, when off N+1 characters, and so on
centering, the x-offset gets rounded.

This commit just adds an extra space after each "On" string so that
it occupies the same width as the "Off" string ... and the when
furiously toggling, the line remains blissfully static.