feat: implement "Create" and "Duplicate" actions
This commit refactors the code from PR #954 to make it work with current
master, as well as only enable the "Create" and "Duplicate" actions when
K9s is not in read-only mode.

Also, both actions will use `kubectl create` instead of `kubectl apply` to
create the resource, as this will automatically prevent the user from overwriting
an existing resource by name.