feat(plugin): add plugin config setting to forward interrupt signals
The following adds a plugin config setting that allows plugins to
define that they would like to receive the Interrupt signal. This
can be usefull in cases where a plugin starts a shell (bash, fish, ...)
and rather want's to handle the interrupt instead of having k9s
exit the process.