Consider readinessGates + ready condition in diagnose
Now that readiness gates are supported, those should be considered when displaying a pod healthiness

Additionally consider the pod ready condition status. It should match the && of the containers' ready condition and readiness gates but I've actually observed that in some cases, it can be false with the containers reporting ready true.